Refactor ZenFlow integration in DeepSpeedEngine
- Move `_configure_zenflow` logic to a standalone `configure_zenflow()` function in `zenflow_utils.py`
- Refactor ZenFlow place to decouple it from ZeRO internals
