--- deepspeed/runtime/zero/zenflow/zenflow_utils.py
@@ -1,42 +1,0 @@
-# Copyright (c) Microsoft Corporation.
-# SPDX-License-Identifier: Apache-2.0
-
-# DeepSpeed Team
-
-import torch
-
-
-def _flatten_dense_tensors(tensors):
-    """Flatten dense tensors into a contiguous 1D buffer. Assume tensors are of
-    same dense type.
-
-    Since inputs are dense, the resulting tensor will be a concatenated 1D
-    buffer. Element-wise operation on this buffer will be equivalent to
-    operating individually.
-
-    Args:
-        tensors (Iterable[Tensor]): dense tensors to flatten.
-
-    Returns:
-        A contiguous 1D buffer containing input tensors.
-    """
-    transposed_tensors = [t.transpose(0, 1).contiguous() if t.dim() == 2 else t for t in tensors]
-    return torch._C._nn.flatten_dense_tensors(transposed_tensors)
-
-
-def _unflatten_dense_tensors(flat, tensors):
-    """View a flat buffer using the sizes of tensors. Assume that tensors are of
-    same dense type, and that flat is given by _flatten_dense_tensors.
-
-    Args:
-        flat (Tensor): flattened dense tensors to unflatten.
-        tensors (Iterable[Tensor]): dense tensors whose sizes will be used to
-          unflatten flat.
-
-    Returns:
-        Unflattened dense tensors with sizes same as tensors and values from
-        flat.
-    """
-    transposed_tensors = [t.transpose(0, 1) if t.dim() == 2 else t for t in tensors]
-    unflat = torch._C._nn.unflatten_dense_tensors(flat, transposed_tensors)
-    return [t.transpose(0, 1) if t.dim() == 2 else t for t in unflat]
